Fix page reload after submitting form & summary calculation
- Page Reload

Event listener for the form was added to the form element itself.
Because the form was rerendered after form submition (together with
summary), there was a  new form element that wasn't listened
for events.
Now the event listener for form submition is attached to the calc
element that cotains form and summary and we select form using
event bubbling.
Additionally seperate view calcViewInputSummary for rendering
summary was created to avoid rerendering of the form.

- Summary calculation

In getDataPoints function the intervalNumber wasn't converted to
a number in getDataPoints function which resulted in adding
huge amount of months or weeks to current date within while loop,
finishing it after one or few iterations even though there
should've been much more..