Add rendering summary only when form was submitted at least once
Add formSubmitted boolean to state to track if form was submitted
so summary doesn't get rendered when form wasn't yet submitted

Fix assigning value to null when other page then calc was clicked.
When, for example, contact page was clicked, there was no summary
container to render summary into which resulted in an  error.
Solution was to add guard clause in calViewInputSummary.js render
function to abort rendering when there's no parentElement to render
into