Change main-nav functionality to listening for load and hashchange
Before the commit, the main-nav and cta button weres based on click
events. Because of that, when user would click back arrow to go to
pevious page  view, it wouldn't work or would work in an unexpected
way.

After commit, main-nav buttons respond naturally as links, generating
hash change, which then triggers rendering of the selected page. That
way, when the user wan't to go back using back arrow in the browser,
he can move to the previous page view. And it's a lot shorter solution.

Additionally, home page code was also moved to generateMarkup method
in mainView so the entire main part of the page content is generated
with javascript.