Change calc nav to render after form is submitted and data fetched
Before calc nav was rendered together with calc view, which allowed
user to go to chart and table views even though form wasn't yet
submitted or new data fetched to be rendered.

Now calc nav will only be visible when form is submitted and data
is fetched. To do that, calc nav buttons are being deleted whenever
user submits the form and rendered back when data is fetched.