Add spinner and error message after form submit on mobile
Form submit button was wrapped in a container same as the container
for spinner, error and back-to-input button in summary.

Class for rendering back-to-input button, spinner and errors
was adjusted to work for both mobile (renders in input part)
and desktop (renders in summary part). This requires to pass
both what user wants to render and in which view part