Fix button-spinner-error bug & Change styles
- Fix:

In controller.js before rendering into button-spinner-error
element we make sure it's proper view: mobile or desktop

In calcViewInput.js there was previous class name to the
button-spinner-error element which prevented renders on mobile
as there was no parent element to render to

- Changes:

Way of displaying pages. Before main content of all pages was
centered, which didn't look good for other pages then home page
and calc. Now only these to mentioned pages are centered.

To achieve that, main element got self-align: start, which made
pages main content to move to the top

To make centering of the home and calc pages possible, when rendering
page we add or remove .center class to the .container element and then
style main element to self-align: center only when there's center class
in the element containing main (.container)