Fix long chart and table render when resizing screen in mobile
When user changed screen width and some of that width
was within mobile width, even if then switched to desktop,
when navigating to chart or table view, it took a long
time for chart or table to render.

The issue was that whenever resize handler was called, it added
event listeners to the calc nav. The more times the handler
was called, the more event listeners were stacked. Then, when
clicking on chart or table,  it was trying to render each
view multiple times creating lag.

Solution was to remove calc nav rendering and adding handlers
from controlMainElementResize function