Fix rendering proper calc view on resize within mobile view
To make it possible previously calcView was addes to state
to keep track of currently selected view. This wasn't enough
as when using controlCalcView function to render calc view
on every resize, it was triggering rerendering even though
the view didn't change from mobile to desktop. To avoid that,
previousMobile view is added to keep track if both current
mobile view is true and if previous mobile view was true
and rerender view only when mobile !=== mobilePrevious