Refactor ButtonSpinnerError rendering into seperate methods
Before, to render button, spinner or error in ButtonSpinnerError
element, controller needed to pass into render method the element
it wanted to render and part of the form it wanted to render to
based on wether in or not in mobile view.

Now controller just uses method with name that tells what it will
render and the render method, pass mobile variable from state and
error message for renderError and rest of the logic around
selecting view part etc is left to calcViewButtonSpinnerError to
handle and render everything properly.

That way we achieve better seperattion of Controller and View