Change _formatNumber method
Before, _formatNumber was formatting number using currency style and
deleting last 4 characters of the resulting string. This, depending
on the user's locale, could delete part of the number, as for some
locale the currency symbol would be in front, and for some at the end
of the number.
Now, using decimal style, there's no currency smybol, just number
formatted according to users locale